Tests/LibWeb/Text/expected/css/animation-name-removal-cancels-animation.txt

with both animations: a,b
after removing 'b' from animation-name: a
